requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
idna==2.8
importlib-metadata==0.18
more-itertools==7.1.0
opencc-python-reimplemented==0.1.4
packaging==19.0
pdfminer3k==1.3.1
pinyin==0.4.0
pluggy==0.12.0
ply==3.11
py==1.8.0
pyparsing==2.4.0
pytest==5.0.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.1
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.5.1
